Step detection uses global vertical acceleration.
Step detection previously checked acceleration in the z axis of the device's frame of reference (into and out of the screen). This did not accurately detect steps when the device was held in orientations other than parallel to the ground. Step detection now checks acceleration in the global z axis frame of reference (into and out of the center of the Earth).